@@ -27,7 +27,7 @@ type EventDispatcher struct {
 func Generate(file File) ([]byte, error) {
 	code := jen.NewFilePathName(file.Package.Path, file.Package.Name)
 
-	code.HeaderComment("// +build !ignore_autogenerated\n")
+	code.HeaderComment("//go:build !ignore_autogenerated\n// +build !ignore_autogenerated\n")
 
 	if file.HeaderText != "" {
 		code.HeaderComment(file.HeaderText)
