@@ -63,7 +63,8 @@ func TestGenerate(t *testing.T) {
 		},
 	}
 
-	expected := `// +build !ignore_autogenerated
+	expected := `//go:build !ignore_autogenerated
+// +build !ignore_autogenerated
 
 // Code generated by mga tool. DO NOT EDIT.
 
