@@ -35,7 +35,8 @@ func TestGenerate(t *testing.T) {
 		},
 	}
 
-	expected := `// +build !ignore_autogenerated
+	expected := `//go:build !ignore_autogenerated
+// +build !ignore_autogenerated
 
 // Copyright 2020 Acme Inc.
 // All rights reserved.
