@@ -55,6 +55,15 @@ func (g Generator) RegisterMarkers(into *markers.Registry) error {
 	return nil
 }
 
+func (Generator) CheckFilter() loader.NodeFilter {
+	return func(node ast.Node) bool {
+		// ignore non-interfaces
+		// _, isIface := node.(*ast.InterfaceType)
+
+		return true
+	}
+}
+
 func (g Generator) Generate(ctx *genall.GenerationContext) error {
 	var headerText string
 
@@ -77,12 +86,7 @@ func (g Generator) Generate(ctx *genall.GenerationContext) error {
 }
 
 func (g Generator) generatePackage(ctx *genall.GenerationContext, headerText string, root *loader.Package) {
-	ctx.Checker.Check(root, func(node ast.Node) bool {
-		// ignore non-interfaces
-		// _, isIface := node.(*ast.InterfaceType)
-
-		return true
-	})
+	ctx.Checker.Check(root)
 
 	root.NeedTypesInfo()
 
